Convert numeric types before multiplication

+++ src/layers/tile/finite.rs
@@ -53,7 +53,7 @@ impl FiniteTileLayerData {
 
     pub(crate) fn get_tile(&self, x: i32, y: i32) -> Option<&LayerTileData> {
         if x < self.width as i32 && y < self.height as i32 && x >= 0 && y >= 0 {
-            self.tiles[(x + y * self.width as i32) as usize].as_ref()
+            self.tiles[x as usize + y as usize * self.width as usize].as_ref()
         } else {
             None
         }
